data-extractor/requirements.txt
bs4
pandas
psycopg2-binary
requests
lxml
